requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
jinja2==3.1.2
numpy==1.24.3
scipy==1.11.3
fastdtw==0.3.4
pandas==2.1.2